refactor(RunImage): split tabs in sub-components
